requirements.txt
python-dotenv
langchain
langchain-openai
PyYAML
rich
